requirements.txt
pytest
pytest-xdist
Appium-Python-Client
python-dotenv